🚀 왼쪽 버튼 눌렀을 때 움직이면 그리기

+++ G24W03MFC/G24W03MFCDoc.h
@@ -12,7 +12,7 @@ protected:
 	//CPoint Point = CPoint(-100, -100);
 	CArray<CPoint, CPoint> Points;
 public:
-	int GetPointsCount() { return Points.GetCount(); }
+	int GetPointsCount() { return (int)Points.GetCount(); }
 
 	//CPoint GetPoint() { return Point; }
 	CPoint GetPoint(int index) { return Points[index]; }
